requirements.txt
flask==3.0.0
pydantic==2.5.2
kubernetes==28.1.0
openai==1.3.5
python-dotenv==1.0.0
httpx==0.24.1 